下载大小和总大小都用长整型

+++ netlib/src/main/java/com/sup/itg/netlib/okhttpLib/Task.java
@@ -118,7 +118,7 @@ public class Task implements ItgTask {
     }
 
     @Override
-    public float getDownloadSize() {
+    public long getDownloadSize() {
         return rDownloadSize;
     }
